GeneticAlgorithm

Documentation for GeneticAlgorithm.

GeneticAlgorithm.binarystring

GeneticAlgorithm.default_selection

GeneticAlgorithm.genAlgo

GeneticAlgorithm.geneticAlgorithm

GeneticAlgorithm.griewank

GeneticAlgorithm.k_point_crossover

GeneticAlgorithm.quartic

GeneticAlgorithm.rastrigin

GeneticAlgorithm.rosenbrock

GeneticAlgorithm.schwefel

GeneticAlgorithm.single_point_crossover

GeneticAlgorithm.solveRosenbrock

GeneticAlgorithm.sphere

GeneticAlgorithm.tournament_selection

GeneticAlgorithm.transformRange

GeneticAlgorithm.uniform_crossover

GeneticAlgorithm.weighted_selection

GeneticAlgorithm.binarystring — Method

binarystring(x)

Compute the fitness of a binary string unit.

Arguments

x: Input vector.

Returns

The digit sum of input vector.

GeneticAlgorithm.default_selection — Method

default_selection(population::Vector, num::Integer)

Select num random genes from population.

Arguments

population: Population vector.

num: Amount of genes to select

Returns

num selected genes

GeneticAlgorithm.genAlgo — Method

genAlgo( popSize::Integer, unitValues::Type, unitShape::Integer, fitnessFunc::Function, genNum::Integer, selection::Function, crossover::Function, mutation::Function, crossRate::Real, mutRate::Real )

TODO: description

Arguments

fitnessFunc: Fitness function.

popSize: Size of population.

unitValues: Type of unit.

unitShape: Length of a unit vector.

genNum: Number of generations.

selection: Selection function.

crossover: Crossover function.

mutation: Mutation function.

crossRate: Crossover rate.

mutRate: Mutation rate.

GeneticAlgorithm.geneticAlgorithm — Method

geneticAlgorithm( popSize::Integer, unitValues::Type, unitShape::Integer, fitnessFunc::Function, genNum::Integer, selection::Function, crossover::Function, mutation::Function, crossRate::Real, mutRate::Real, nextGenAmt::Real, terminationNum::Real; initFunc::Function )

Evolve the population of units for a specified number of generations. The loop creates a new population every generation by selecting units with higher fitness, recombinating and mutating them until the old population can be replaced by a fitter one.

Arguments

fitnessFunc: Fitness function.

popSize: Size of population.

unitValues: Type of unit.

unitShape: Length of a unit vector.

genNum: Number of generations.

selection: Selection function.

crossover: Crossover function.

mutation: Mutation function.

crossRate: Crossover rate.

mutRate: Mutation rate.

nextGenAmt: amount of genes that are automatically copied to new generation

terminationNum: number of iteration after which the algorithm aborts if there is no change in fitness value from the best gene

initFunc: function to generate initial population

GeneticAlgorithm.griewank — Method

griewank(x)

Compute the Griewank function value for a given input vector x.

The Griewank function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as the difference between two terms: the sum of the squares of each element in x divided by 4000, and the product of the cosine of each element in x divided by the square root of its index.

Arguments

x: Input vector.

Returns

The Griewank function value.

GeneticAlgorithm.k_point_crossover — Method

k_point_crossover(gene1::Union{AbstractVector, AbstractMatrix}, gene2::Union{AbstractVector, AbstractMatrix}, k::Integer)

Recombinate two units by exchanging their genes at k random points

Arguments

gene1: Parent unit 1.

gene2: Parent unit 2.

k: # of crossover points.

Returns

Two recombined child units.

GeneticAlgorithm.quartic — Method

quartic(x)

Compute the quartic function value for a given input vector x.

The quartic function is defined as the sum of the product of each element in x raised to the power of 4 and its index, plus a random number.

Arguments

x: Input vector.

Returns

The quartic function value.

GeneticAlgorithm.rastrigin — Method

rastrigin(x)

Compute the Rastrigin function value for a given input vector x.

The Rastrigin function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as a sum of terms involving the square of each element in x, minus 10 times the cosine of 2π times each element in x.

Arguments

x: Input vector.

Returns

The Rastrigin function value.

GeneticAlgorithm.rosenbrock — Method

rosenbrock(x)

Compute the Rosenbrock function value for a given input vector x.

The Rosenbrock function is a non-convex function used as a performance test problem for optimization algorithms. It is defined as the sum of a series of terms involving the squares of differences between adjacent elements of x.

Arguments

x: Input vector.

a: value for the variable a

b: value for the variable b

Returns

The Rosenbrock function value.

GeneticAlgorithm.schwefel — Method

schwefel(x)

Compute the Schwefel function value for a given input vector x.

The Schwefel function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as a sum of terms involving the sine function and the square root of the absolute value of each element in x.

Arguments

x: Input vector.

Returns

The Schwefel function value.

GeneticAlgorithm.single_point_crossover — Method

single_point_crossover(gene1::Union{AbstractVector, AbstractMatrix}, gene2::Union{AbstractVector, AbstractMatrix})

Recombinate two units by exchanging their genes from a random index onward.

Arguments

gene1: Parent unit 1.

gene2: Parent unit 2.

Returns

Two recombined child units.

GeneticAlgorithm.solveRosenbrock — Method

solveRosenbrock(; a::Integer, b::Integer, popSize::Integer, fitnessFunc::Function, unitValues::Union{Type, AbstractVector{<:AbstractFloat}, AbstractRange{<:Real}}, unitShape::AbstractVector{<:Integer}, genNum::Integer, crossRate::Real, mutRate::Real, nextGenAmt::Number, selectionFunc::Function, crossoverFunc::Function, mutationFunc::Function , terminationNum::Real )

Runs the rosenbrock function with the geneticAlgorithm

Arguments

a: rosenbrock function variable

b: rosenbrock function variable

popSize: Size of population.

fitnessFunc: Fitness function.

unitValues: Type of unit.

unitShape: Length of a unit vector.

genNum: Number of generations.

crossRate: Crossover rate.

mutRate: Mutation rate.

nextGenAmt: amount of genes that are automatically copied to new generation

selectionFunc: Selection function.

crossoverFunc: Crossover function.

mutationFunc: Mutation function.

terminationNum: number of iteration after which the algorithm aborts if there is no change in fitness value from the best gene

GeneticAlgorithm.sphere — Method

sphere(x)

Compute the sphere function value for a given input vector x.

The sphere function is defined as the sum of the squares of each element in x.

Arguments

x: Input vector.

Returns

The sphere function value.

GeneticAlgorithm.tournament_selection — Method

tournament_selection(population::Vector, fitness::Vector, tournamentSize::Int, num::Integer)

Select one unit with higher fitness over several tournament rounds with randomly chosen groups of unit from the population .

Arguments

population: Population vector.

fitness: Vector of fitness values.

tournamentSize: Size of tournament groups.

num: Amount of genes to select

Returns

The unit with better fitness value.

GeneticAlgorithm.transformRange — Method

transformRange( values::Union{Vector,Matrix}, oldMax::Real, oldMin::Real, newMax::Real, newMin::Real )

Transforms values in a vector or matrix to a new range.

Arguments

'values' Values to be transformed 'oldMax' The maximum value of the old range 'oldMin' The minimum value of the old range 'newMax' The maximum value of the new range 'newMin' The minimum value of the new range

GeneticAlgorithm.uniform_crossover — Method

uniform_crossover(gene1::Union{AbstractVector, AbstractMatrix}, gene2::Union{AbstractVector, AbstractMatrix})

Recombinate two units by iterating a gene and swapping the values with a 50% chance

Arguments

gene1: Parent unit 1.

gene2: Parent unit 2.

Returns

Two recombined child units.

GeneticAlgorithm.weighted_selection — Method

weighted_selection(population::Vector, fitness::Vector, num::Integer)

Select num random genes from population weight by the fitness of the genes

Arguments

population: Population vector.

fitness: Vector of fitness values.

num: Amount of genes to select

Returns

num selected genes