GeneticAlgorithm

Documentation for GeneticAlgorithm.

GeneticAlgorithm.binarystring

GeneticAlgorithm.crossover

GeneticAlgorithm.genAlgo

GeneticAlgorithm.griewank

GeneticAlgorithm.initPop

GeneticAlgorithm.mutation

GeneticAlgorithm.quartic

GeneticAlgorithm.rastrigin

GeneticAlgorithm.rosenbrock

GeneticAlgorithm.schwefel

GeneticAlgorithm.selection

GeneticAlgorithm.sphere

GeneticAlgorithm.tournamentSelection

GeneticAlgorithm.binarystring — Method

binarystring(x)

Compute the fitness of a binary string unit.

Arguments

x: Input vector.

Returns

The digit sum of input vector.

GeneticAlgorithm.crossover — Method

crossover(parent1::Vector, parent2::Vector)

Recombinate two units by exchanging their genes from a random index onward.

Arguments

parent1: Parent unit 1.

parent2: Parent unit 2.

Returns

Two recombined child units.

GeneticAlgorithm.genAlgo — Method

genAlgo(
    popSize::Int,
    unitReal::Bool,
    unitLen::Int,
    fitnessFunc::Function,
    genNum::Int,
    crossRate::Float64,
    mutRate::Float64
)

Evolve the population of units for a specified number of generations. The loop creates a new population every generation by selecting two units with higher fitness, recombinating and mutating them until the old population can be replaced by a fitter one.

Arguments

popSize: Size of population.

unitReal: Binary or real unit representation.

unitLen: Length of a unit vector.

fitnessFunc: Fitness function.

genNum: Number of generations.

crossRate: Crossover rate.

mutRate: Mutation rate.

GeneticAlgorithm.griewank — Method

griewank(x)

Compute the Griewank function value for a given input vector x.

The Griewank function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as the difference between two terms: the sum of the squares of each element in x divided by 4000, and the product of the cosine of each element in x divided by the square root of its index.

Arguments

x: Input vector.

Returns

The Griewank function value.

GeneticAlgorithm.initPop — Method

initPop(popSize::Int, unitReal::Bool, unitLen::Int)

Initialize the population of units. Each unit is a vector of Boolean or Real values.

Arguments

popSize Size of population.

unitReal Binary or real unit representation.

unitLen Length of a unit vector.

Returns

Full vector of units with random Boolean or Real values.

GeneticAlgorithm.mutation — Method

mutation!(unit::Vector, unitReal::Bool, mutRate::Float64)

Mutate an unit by changing its genes depending on its type and a mutation rate.

Arguments

unit: Unit to be mutated.

unitReal: Type of unit, Real or Binary.

mutRate: Mutation rate.

Returns

A mutated unit.

GeneticAlgorithm.quartic — Method

quartic(x)

Compute the quartic function value for a given input vector x.

The quartic function is defined as the sum of the product of each element in x raised to the power of 4 and its index, plus a random number.

Arguments

x: Input vector.

Returns

The quartic function value.

GeneticAlgorithm.rastrigin — Method

rastrigin(x)

Compute the Rastrigin function value for a given input vector x.

The Rastrigin function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as a sum of terms involving the square of each element in x, minus 10 times the cosine of 2π times each element in x.

Arguments

x: Input vector.

Returns

The Rastrigin function value.

GeneticAlgorithm.rosenbrock — Method

rosenbrock(x)

Compute the Rosenbrock function value for a given input vector x.

The Rosenbrock function is a non-convex function used as a performance test problem for optimization algorithms. It is defined as the sum of a series of terms involving the squares of differences between adjacent elements of x.

Arguments

x: Input vector.

Returns

The Rosenbrock function value.

GeneticAlgorithm.schwefel — Method

schwefel(x)

Compute the Schwefel function value for a given input vector x.

The Schwefel function is a multimodal function used as a performance test problem for optimization algorithms. It is defined as a sum of terms involving the sine function and the square root of the absolute value of each element in x.

Arguments

x: Input vector.

Returns

The Schwefel function value.

GeneticAlgorithm.selection — Method

selection(population::Vector, fitness::Vector)

Select one unit with higher fitness out of two randomly chosen unit from the population.

Arguments

population: Population vector.

fitness: Vector of fitness values.

Returns

The unit with better fitness value.

GeneticAlgorithm.sphere — Method

sphere(x)

Compute the sphere function value for a given input vector x.

The sphere function is defined as the sum of the squares of each element in x.

Arguments

x: Input vector.

Returns

The sphere function value.

GeneticAlgorithm.tournamentSelection — Method

tournamentSelection(population::Vector, fitness::Vector, tournamentSize::Int)

Select one unit with higher fitness over several tournament rounds with randomly chosen groups of unit from the population .

Arguments

population: Population vector.

fitness: Vector of fitness values.

tournamentSize: Size of tournament groups.

Returns

The unit with better fitness value.